+++ src/dataflow_spec.py
@@ -25,12 +25,13 @@ class BronzeDataflowSpec:
     readerConfigOptions: map
     targetFormat: str
     targetDetails: map
-    writerConfigOptions: map
+    tableProperties: map
     schema: str
     partitionColumns: list
     cdcApplyChanges: str
     dataQualityExpectations: str
     quarantineTargetDetails: map
+    quarantineTableProperties: map
     version: str
     createDate: datetime
     createdBy: str
@@ -49,7 +50,7 @@ class SilverDataflowSpec:
     readerConfigOptions: map
     targetFormat: str
     targetDetails: map
-    writerConfigOptions: map
+    tableProperties: map
     selectExp: list
     whereClause: list
     partitionColumns: list
@@ -69,11 +70,13 @@ class CDCApplyChanges:
     sequence_by: str
     where: str
     ignore_null_updates: bool
-    apply_as_deletes: bool
-    apply_as_truncates: bool
+    apply_as_deletes: str
+    apply_as_truncates: str
     column_list: list
     except_column_list: list
     scd_type: str
+    track_history_column_list: list
+    track_history_except_column_list: list
 
 
 class DataflowSpecUtils:
@@ -90,6 +93,9 @@ class DataflowSpecUtils:
         "column_list",
         "except_column_list",
         "scd_type",
+        "track_history_column_list",
+        "track_history_except_column_list"
+
     ]
     cdc_applychanges_api_attributes_defaults = {
         "where": None,
@@ -98,6 +104,9 @@ class DataflowSpecUtils:
         "apply_as_truncates": None,
         "column_list": None,
         "except_column_list": None,
+        "track_history_column_list": None,
+        "track_history_except_column_list": None
+
     }
 
     @staticmethod