+++ tests/test_pipeline_readers.py
@@ -14,6 +14,8 @@ import pyspark.sql.types as T
 
 dbutils = MagicMock()
 DBUtils = MagicMock()
+spark = MagicMock()
+spark.readStream = MagicMock()
 
 
 class PipelineReadersTests(DLTFrameworkTestCase):
@@ -28,12 +30,13 @@ class PipelineReadersTests(DLTFrameworkTestCase):
         },
         "targetFormat": "delta",
         "targetDetails": {"database": "bronze", "table": "customer", "path": "tests/localtest/delta/customers"},
-        "writerConfigOptions": {},
+        "tableProperties": {},
         "schema": None,
         "partitionColumns": [""],
         "cdcApplyChanges": None,
         "dataQualityExpectations": None,
         "quarantineTargetDetails": None,
+        "quarantineTableProperties": None,
         "version": "v1",
         "createDate": datetime.now,
         "createdBy": "dlt-meta-unittest",
@@ -64,12 +67,13 @@ class PipelineReadersTests(DLTFrameworkTestCase):
         },
         "targetFormat": "delta",
         "targetDetails": {"database": "bronze", "table": "customer", "path": "tests/localtest/delta/customers"},
-        "writerConfigOptions": {},
+        "tableProperties": {},
         "schema": None,
         "partitionColumns": [""],
         "cdcApplyChanges": None,
         "dataQualityExpectations": None,
         "quarantineTargetDetails": None,
+        "quarantineTableProperties": None,
         "version": "v1",
         "createDate": datetime.now,
         "createdBy": "dlt-meta-unittest",
@@ -79,7 +83,7 @@ class PipelineReadersTests(DLTFrameworkTestCase):
     bronze_kafka_dataflow_spec_map = {
         "dataFlowId": "1",
         "dataFlowGroup": "A1",
-        "sourceFormat": "eventhub",
+        "sourceFormat": "kafka",
         "sourceDetails": {
             "source_schema_path": "tests/resources/schema/eventhub_iot_schema.ddl",
             "subscribe": "iot",
@@ -91,12 +95,13 @@ class PipelineReadersTests(DLTFrameworkTestCase):
         },
         "targetFormat": "delta",
         "targetDetails": {"database": "bronze", "table": "customer", "path": "tests/localtest/delta/customers"},
-        "writerConfigOptions": {},
+        "tableProperties": {},
         "schema": None,
         "partitionColumns": [""],
         "cdcApplyChanges": None,
         "dataQualityExpectations": None,
         "quarantineTargetDetails": None,
+        "quarantineTableProperties": None,
         "version": "v1",
         "createDate": datetime.now,
         "createdBy": "dlt-meta-unittest",
@@ -136,6 +141,30 @@ class PipelineReadersTests(DLTFrameworkTestCase):
         customer_df = PipelineReaders.read_dlt_cloud_files(self.spark, bronze_dataflow_spec, schema)
         self.assertIsNotNone(customer_df)
 
+    def test_read_delta_positive(self):
+        """Test read_cloud_files positive."""
+        bronze_map = PipelineReadersTests.bronze_dataflow_spec_map
+        source_format_map = {"sourceFormat": "delta"}
+        bronze_map.update(source_format_map)
+        source_details_map = {"sourceDetails": {"path": "tests/resources/delta/customers"}}
+        bronze_map.update(source_details_map)
+        bronze_dataflow_spec = BronzeDataflowSpec(**bronze_map)
+        customer_df = PipelineReaders.read_dlt_delta(self.spark, bronze_dataflow_spec)
+        self.assertIsNotNone(customer_df)
+
+    def test_read_delta_with_read_config_positive(self):
+        """Test read_cloud_files positive."""
+        bronze_map = PipelineReadersTests.bronze_dataflow_spec_map
+        source_format_map = {"sourceFormat": "delta"}
+        bronze_map.update(source_format_map)
+        source_details_map = {"sourceDetails": {"path": "tests/resources/delta/customers"}}
+        bronze_map.update(source_details_map)
+        reader_config = {"readerConfigOptions": {"maxFilesPerTrigger": "1"}}
+        bronze_map.update(reader_config)
+        bronze_dataflow_spec = BronzeDataflowSpec(**bronze_map)
+        customer_df = PipelineReaders.read_dlt_delta(self.spark, bronze_dataflow_spec)
+        self.assertIsNotNone(customer_df)
+
     @patch.object(PipelineReaders, "get_db_utils", return_value=dbutils)
     @patch.object(dbutils, "secrets.get", return_value={"called"})
     def test_get_eventhub_kafka_options(self, get_db_utils, dbutils):
@@ -154,7 +183,7 @@ class PipelineReadersTests(DLTFrameworkTestCase):
         source_details_map = {
             **source_details,
             "kafka.ssl.truststore.location": "tmp:/location",
-            "kafka.ssl.keystore.location": "tmp:/location"
+            "kafka.ssl.keystore.location": "tmp:/location",
         }
         bronze_map['sourceDetails'] = source_details_map
         bronze_dataflow_spec = BronzeDataflowSpec(**bronze_map)
@@ -172,3 +201,31 @@ class PipelineReadersTests(DLTFrameworkTestCase):
         """Test Get kafka options."""
         dbutils = PipelineReaders.get_db_utils(self.spark)
         self.assertIsNotNone(dbutils)
+
+    @patch.object(spark, "readStream", return_value={"called"})
+    @patch.object(dbutils, "secrets.get", return_value={"called"})
+    def test_kafka_positive(self, spark, dbutils):
+        """Test kafka read positive."""
+        bronze_map = PipelineReadersTests.bronze_kafka_dataflow_spec_map
+        source_details = bronze_map['sourceDetails']
+        source_details_map = {
+            **source_details,
+            "kafka.ssl.truststore.location": "tmp:/location",
+            "kafka.ssl.keystore.location": "tmp:/location",
+            "kafka.ssl.truststore.secrets.scope": "databricks",
+            "kafka.ssl.truststore.secrets.key": "databricks",
+            "kafka.ssl.keystore.secrets.scope": "databricks",
+            "kafka.ssl.keystore.secrets.key": "databricks"
+        }
+        bronze_map['sourceDetails'] = source_details_map
+        bronze_dataflow_spec = BronzeDataflowSpec(**bronze_map)
+        customer_df = PipelineReaders.read_kafka(spark, bronze_dataflow_spec, None)
+        self.assertIsNotNone(customer_df)
+
+    @patch.object(spark, "readStream", return_value={"called"})
+    def test_eventhub_positive(self, spark):
+        """Test eventhub read positive."""
+        bronze_map = PipelineReadersTests.bronze_eventhub_dataflow_spec_map
+        bronze_dataflow_spec = BronzeDataflowSpec(**bronze_map)
+        customer_df = PipelineReaders.read_kafka(spark, bronze_dataflow_spec, None)
+        self.assertIsNotNone(customer_df)