+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 **NOTE:** For CLI interfaces, we support SemVer approach. However, for API components we don't use SemVer as of now. This may lead to instability when using dbx API methods directly.
 
 [Please read through the Keep a Changelog (~5min)](https://keepachangelog.com/en/1.0.0/).
+## [v0.0.2] - 2023-05-11
+### Added
+- Table properties support for bronze, quarantine and silver tables using create_streaming_live_table api call
+- Support for track history column using apply_changes api
+- Support for delta as source
+- Validation for bronze/silver onboarding
+### Fixed
+- Input schema parsing issue in onboarding
+### Modified
+-  Readme and docs to include above features
 
 ## [v0.0.1] - 2023-03-22
 ### Added